refactor(game-session): split ActiveGameSession concerns
Pulls side effects into a hook and breaks the UI into
narrative, choice, and control sections, which keeps the core
component easier to follow.

Adds a typed auto-save return for reuse in the new controls.

Closes #989
