+++ src/components/GameSession/hooks/useActiveGameSessionEffects.ts
@@ -1,0 +1,283 @@
+'use client';
+
+import { useCallback, useEffect } from 'react';
+import type { Dispatch, MutableRefObject, SetStateAction } from 'react';
+import type { Decision, NarrativeSegment } from '@/types/narrative.types';
+import { useNarrativeStore } from '@/state/narrativeStore';
+import { useSessionStore } from '@/state/sessionStore';
+
+const INITIAL_GENERATION_MAX_WAIT_MS = 20000;
+const CHOICE_FALLBACK_DELAY_MS = 15000;
+
+interface UseActiveGameSessionEffectsOptions {
+  sessionId: string;
+  worldId: string;
+  controllerKey: string;
+  initialized: boolean;
+  isGenerating: boolean;
+  segmentCount: number;
+  characterId?: string;
+  onEnd?: () => void;
+  onEndStoryClick: () => void;
+  setIsGenerating: Dispatch<SetStateAction<boolean>>;
+  setInitialized: Dispatch<SetStateAction<boolean>>;
+  setCurrentDecision: Dispatch<SetStateAction<Decision | null>>;
+  setIsGeneratingChoices: Dispatch<SetStateAction<boolean>>;
+  setIsFatalEnding: Dispatch<SetStateAction<boolean>>;
+  fatalEndingTriggeredRef: MutableRefObject<boolean>;
+  choiceGenerationTimeoutRef: MutableRefObject<NodeJS.Timeout | null>;
+}
+
+export const useActiveGameSessionEffects = ({
+  sessionId,
+  worldId,
+  controllerKey,
+  initialized,
+  isGenerating,
+  segmentCount,
+  characterId,
+  onEnd,
+  onEndStoryClick,
+  setIsGenerating,
+  setInitialized,
+  setCurrentDecision,
+  setIsGeneratingChoices,
+  setIsFatalEnding,
+  fatalEndingTriggeredRef,
+  choiceGenerationTimeoutRef,
+}: UseActiveGameSessionEffectsOptions) => {
+  // Reset fatal guard and flag when session changes
+  useEffect(() => {
+    fatalEndingTriggeredRef.current = false;
+    setIsFatalEnding(false);
+  }, [sessionId, setIsFatalEnding, fatalEndingTriggeredRef]);
+
+  // Safety net: if no narrative segment arrives within a reasonable window,
+  // inject a minimal fallback scene so the UI can progress.
+  // Only trigger if we're not actively generating content.
+  useEffect(() => {
+    if (!initialized) return;
+    if (segmentCount > 0) return;
+    let cancelled = false;
+    const t = setTimeout(() => {
+      if (cancelled) return;
+      // Don't inject fallback if AI generation is still in progress
+      if (isGenerating) return;
+
+      try {
+        const now = new Date();
+        const fallback: NarrativeSegment = {
+          id: `seg-${sessionId}-bootstrap-${now.getTime()}`,
+          content: 'You take a breath as your adventure begins. The world awaits your first move.',
+          type: 'scene',
+          metadata: { location: 'Starting Location', tags: ['intro', 'bootstrap'] },
+          sessionId,
+          worldId,
+          timestamp: now,
+          createdAt: now.toISOString(),
+          updatedAt: now.toISOString(),
+        } as NarrativeSegment;
+        // Add to store only; NarrativeHistoryManager reads from store
+        useNarrativeStore.getState().addSegment(sessionId, {
+          content: fallback.content,
+          type: fallback.type,
+          characterIds: [],
+          metadata: fallback.metadata,
+          updatedAt: fallback.updatedAt,
+          timestamp: fallback.timestamp,
+        });
+        // Begin generating choices after bootstrap
+        setIsGeneratingChoices(true);
+      } catch {
+        // Ignore errors; controller may be mid-flight
+      }
+    }, 4000); // Increased timeout to 4 seconds to give AI more time
+    return () => {
+      cancelled = true;
+      clearTimeout(t);
+    };
+  }, [initialized, segmentCount, sessionId, worldId, isGenerating, setIsGeneratingChoices]);
+
+  // Ensure we eventually release the generating flag to allow safety fallbacks
+  useEffect(() => {
+    if (!initialized) return;
+    if (segmentCount > 0) return;
+    if (!isGenerating) return;
+
+    let cancelled = false;
+    const timeoutId = setTimeout(() => {
+      if (cancelled) return;
+      setIsGenerating(false);
+    }, INITIAL_GENERATION_MAX_WAIT_MS);
+
+    return () => {
+      cancelled = true;
+      clearTimeout(timeoutId);
+    };
+  }, [initialized, segmentCount, isGenerating, setIsGenerating]);
+
+  // Initialize the narrative only once per session
+  // instead of clearing and recreating each time
+  useEffect(() => {
+    // Initialize session with unique controller key
+    let isMounted = true;
+
+    // Set initial loading state
+    setIsGenerating(true);
+
+    // Function to check existing narrative and set up if needed
+    const setupNarrative = async () => {
+      try {
+        // Dynamically import the narrativeStore to avoid circular dependencies
+        const { useNarrativeStore } = await import('@/state/narrativeStore');
+
+        // Only proceed if still mounted
+        if (!isMounted) return;
+
+        // Check if we already have segments for this session
+        const existingSegments = useNarrativeStore.getState().getSessionSegments(sessionId);
+        // Check for 'intro' tag which is more stable than checking specific location strings
+        const hasInitialScene = existingSegments.some(seg =>
+          seg.metadata?.tags?.includes('intro')
+        );
+
+        // Check for existing decisions in the store
+        const existingDecisions = useNarrativeStore.getState().getSessionDecisions(sessionId);
+
+        // If we have existing decisions, use the latest one
+        if (existingDecisions.length > 0) {
+          const latestDecision = existingDecisions[existingDecisions.length - 1];
+          setCurrentDecision(latestDecision);
+        }
+
+        if (hasInitialScene || existingSegments.length > 0) {
+          // If we have any segments at all, use them
+          // Don't clear existing narrative history
+          setInitialized(true);
+          setIsGenerating(false);
+          // Choice generation will be triggered by NarrativeController after narrative generation
+        }
+        else {
+          // No segments at all - normal case for new session
+          // Keep UI in generating state until first segment arrives or we explicitly fallback
+          setInitialized(true);
+          setIsGenerating(true);
+        }
+      } catch {
+        // Error setting up narrative, continue with initialization
+        setInitialized(true);
+        setIsGenerating(true);
+      }
+    };
+
+    // Check existing narrative and set up if needed
+    setupNarrative();
+
+    return () => {
+      // Mark component as unmounted to prevent state updates after unmounting
+      isMounted = false;
+
+      // Clear any pending choice generation timeout
+      if (choiceGenerationTimeoutRef.current) {
+        clearTimeout(choiceGenerationTimeoutRef.current);
+        choiceGenerationTimeoutRef.current = null;
+      }
+    };
+  }, [sessionId, worldId, controllerKey, setIsGenerating, setInitialized, setCurrentDecision, choiceGenerationTimeoutRef]);
+
+  // Keep currentDecision synchronized with store updates (supports external generators like item usage)
+  useEffect(() => {
+    const unsubscribe = useNarrativeStore.subscribe((state) => {
+      const decisionIds = state.sessionDecisions[sessionId];
+      const ids = decisionIds || [];
+      const latestId = ids[ids.length - 1];
+      if (latestId) {
+        const latest = state.decisions[latestId] || null;
+        setCurrentDecision(latest);
+        setIsGeneratingChoices(false);
+      } else {
+        setCurrentDecision(null);
+        // If narrative already exists, surface a loading state while choices regenerate
+        const hasSegments = (state.sessionSegments[sessionId]?.length ?? 0) > 0;
+        if (hasSegments) {
+          setIsGeneratingChoices(true);
+        }
+      }
+    });
+
+    return () => {
+      unsubscribe();
+    };
+  }, [sessionId, setCurrentDecision, setIsGeneratingChoices]);
+
+  // Global event handlers to support hero action buttons from parent pages
+  useEffect(() => {
+    const onEndStory = () => onEndStoryClick();
+    const onEndSession = async () => {
+      // Dispatch event to trigger final checkpoint before ending session
+      window.dispatchEvent(new CustomEvent('narraitor:finalize-checkpoint'));
+      // Small delay to allow checkpoint to complete
+      await new Promise(resolve => setTimeout(resolve, 500));
+      if (onEnd) onEnd();
+    };
+    const onNewSession = async () => {
+      const sessionStore = useSessionStore.getState();
+      const narrativeStore = useNarrativeStore.getState();
+      narrativeStore.clearSessionSegments(sessionId);
+      narrativeStore.clearSessionDecisions(sessionId);
+      narrativeStore.clearEnding();
+      sessionStore.endSession();
+      Object.keys(sessionStore.savedSessions).forEach(savedSessionId => {
+        const savedSession = sessionStore.savedSessions[savedSessionId];
+        if (savedSession.worldId === worldId && savedSession.characterId === characterId) {
+          sessionStore.deleteSavedSession(savedSessionId);
+        }
+      });
+      await new Promise(resolve => setTimeout(resolve, 100));
+      const url = new URL(window.location.href);
+      url.searchParams.set('fresh', 'true');
+      window.location.href = url.toString();
+    };
+
+    window.addEventListener('narraitor:end-story', onEndStory as EventListener);
+    window.addEventListener('narraitor:end-session', onEndSession as EventListener);
+    window.addEventListener('narraitor:new-session', onNewSession as EventListener);
+    return () => {
+      window.removeEventListener('narraitor:end-story', onEndStory as EventListener);
+      window.removeEventListener('narraitor:end-session', onEndSession as EventListener);
+      window.removeEventListener('narraitor:new-session', onNewSession as EventListener);
+    };
+  }, [sessionId, worldId, characterId, onEnd, onEndStoryClick]);
+
+  const scheduleChoiceFallback = useCallback((decisionAtSchedule: Decision | null) => {
+    const timeoutId = setTimeout(() => {
+      // If we're still generating choices after delay, create fallback choices
+      setIsGeneratingChoices(prev => {
+        if (prev && !decisionAtSchedule) {
+          const fallbackId = `decision-timeout-${Date.now()}`;
+          const fallbackDecision: Decision = {
+            id: fallbackId,
+            prompt: 'What will you do?',
+            options: [
+              { id: `option-${fallbackId}-1`, text: 'Investigate further', alignment: 'neutral' },
+              { id: `option-${fallbackId}-2`, text: 'Talk to nearby characters', alignment: 'lawful' },
+              { id: `option-${fallbackId}-3`, text: 'Move to a new location', alignment: 'neutral' },
+            ],
+            decisionWeight: 'minor',
+            contextSummary: 'Waiting for player action (timeout fallback).',
+          };
+
+          setCurrentDecision(fallbackDecision);
+          return false; // Stop generating
+        }
+        return prev;
+      });
+    }, CHOICE_FALLBACK_DELAY_MS);
+
+    choiceGenerationTimeoutRef.current = timeoutId;
+  }, [choiceGenerationTimeoutRef, setCurrentDecision, setIsGeneratingChoices]);
+
+  return {
+    scheduleChoiceFallback,
+  };
+};
